Added control for pause button
Upon pressing the button(connected to digital pin 1, as pin ~2 fail to transmit the signal), the worm relax to {0,0,0,0,0,0,0} state. The following press would make it contract again. This prevents having to perform physical recalibration too often.

Additionally, fix the memory leak issue in Routine.cpp where an extra segment is accessed.

Need to update this mechanism to the undulation function.